translate - Show your support (*)
* note that "Full de trucs" has been changed to "Apunts"
The term "cheat sheet" has no direct translation in Catalan. In Spanish, it should have been translated to "chuleta", which has the colloquial translation "xuleta" in Catalan, but currently not accepted in the Catalan dictionary. "Full de trucs" was a (bad) literal translation of the English version because I've always used to "cheat sheet" or "xuleta" to refer to that term.